pyproject.toml
[project]
name = "fast-mcp-agent"
version = "0.1.0"
description = "Multi-agent AI system powered by xAI Grok, with Slack UI, web research, browser automation, Google Workspace, and blockchain capabilities."
requires-python = ">=3.11"
dependencies = [
    "fastmcp[tasks]>=3.0.0b1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "py-key-value-aio[redis]",
    "asyncpg>=0.29.0",
    "neon-api>=0.3.0",
    # xAI Python SDK (gRPC-based, replaces OpenRouter)
    "xai-sdk>=1.0.0",
    # Slack integration (Bolt for inbound events + SDK for outbound API)
    "slack-sdk>=3.33.0",
    "slack-bolt>=1.21.0",
    # Cron scheduling
    "apscheduler>=3.10.0",
    "cron-descriptor>=1.4.0",
    # Token counting (fast local fallback for compaction checks)
    "tiktoken>=0.8.0",
    # Coinbase AgentKit for blockchain operations (optional at runtime)
    "coinbase-agentkit>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
otel = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fast_mcp_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
